Split MiqAeWorkspace and MiqAeWorkspaceRuntime into their own files
